requirements.txt
fastapi[all]
uvicorn[standard]
gunicorn
pydantic
google-generativeai
pytest
pytest-mock
pytest-asyncio
httpx
google-auth-oauthlib
google-api-python-client
itsdangerous
